feat: 주문생성 (동시성 미적용)

+++ src/main/java/com/example/jjangushrine/domain/cart/dto/request/CartItemCreateReq.java
@@ -6,6 +6,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record CartItemCreateReq(
         @JsonProperty("productId") Long productId,
-        @JsonProperty("quantity") int quantity
+        @JsonProperty("quantity") Short quantity
 ) {
 }
