+++ src/main/java/com/example/jjangushrine/domain/cart/dto/response/CartItemCreateRes.java
@@ -1,5 +1,5 @@
 package com.example.jjangushrine.domain.cart.dto.response;
 
 
-public record CartItemCreateRes(String cartId, Long productId, String productName, int quantity, int totalPrice) {
+public record CartItemCreateRes(String cartId, Long productId, String productName, Short quantity, int totalPrice) {
 }
